requirements.txt
streamlit==1.31.0
SpeechRecognition==3.10.0
gTTS==2.5.0
google-generativeai==0.3.2
pytz==2024.1
PyAudio==0.2.14
protobuf~=4.25.5
reportlab~=4.2.5